use existing prettier config when available
