+++ generators/app/index.js
@@ -22,9 +22,6 @@ const BaseGenerator = require('../generator-base');
 const cleanup = require('../cleanup');
 const prompts = require('./prompts');
 const packagejs = require('../../package.json');
-const through = require('through2');
-const filter = require('gulp-filter');
-const prettier = require('prettier');
 
 module.exports = class extends BaseGenerator {
     constructor(args, opts) {
@@ -129,6 +126,7 @@ module.exports = class extends BaseGenerator {
         this.useYarn = this.configOptions.useYarn = !this.options.npm;
         this.isDebugEnabled = this.configOptions.isDebugEnabled = this.options.debug;
         this.experimental = this.configOptions.experimental = this.options.experimental;
+        this.registerClientTransforms();
     }
 
     get initializing() {
@@ -206,41 +204,6 @@ module.exports = class extends BaseGenerator {
                         this.clientPackageManager = 'npm';
                     }
                 }
-            },
-
-            registerPrettier() {
-                if (this.skipClient) {
-                    return;
-                }
-                const tsFilter = filter(['**/*.{ts,tsx}'], { restore: true });
-
-                const tsPrettier = function () {
-                    function transform(file, encoding, callback) {
-                        const str = file.contents.toString('utf8');
-                        const options = {
-                            printWidth: 140,
-                            singleQuote: true,
-                            jsxBracketSameLine: false,
-                            parser: 'typescript',
-                            arrowParens: 'always',
-                            // for better errors
-                            filepath: file.relative
-                        };
-                        const data = prettier.format(str, options);
-                        file.contents = Buffer.from(data);
-
-                        callback(null, file);
-                    }
-
-                    return through.obj(transform);
-                };
-
-                // this pipe will pass through (restore) anything that doesn't match tsFilter
-                this.registerTransformStream([
-                    tsFilter,
-                    tsPrettier(),
-                    tsFilter.restore
-                ]);
             }
         };
     }
