+++ generators/client/templates/angular/.prettierrc.ejs
@@ -3,4 +3,4 @@ printWidth: 140
 singleQuote: true
 jsxBracketSameLine: false
 parser: typescript
-arrowParens: always
+arrowParens: avoid
