+++ package.json
@@ -86,8 +86,8 @@
     "yeoman-test": "1.7.0"
   },
   "engines": {
-    "node": ">=6.9.0",
-    "npm": ">=2.14.2"
+    "node": ">=8.9.0",
+    "npm": ">=5.5.1"
   },
   "license": "Apache-2.0",
   "collective": {
